requirements.txt
celery
dj_database_url
django
python-dotenv
djangorestframework
python-telegram-bot
pytest-cov
django-cors-headers
psycopg2-binary
coreapi
flake8
djangorestframework-simplejwt
drf-yasg
drf-spectacular
